Restructure robot module and add connect handling in worker
