+++ app/ui/robot/name.js
@@ -1,0 +1,4 @@
+// robot redux module name constant
+// TODO(mc, 2017-08-29): maybe do a package.json instead?
+
+export default 'robot'
